requirements.txt
asgiref==3.8.1
autopep8==2.3.2
certifi==2024.8.30
Django==5.0.7
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
generics==7.0.0
gprof2dot==2024.6.6
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.12.1
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.1
sentry-sdk==2.18.0
setuptools==72.1.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
